requirements.txt
flask>=3.0.0
firebase-admin>=6.0.0
python-dotenv>=1.0.0
werkzeug>=3.0.0
numpy>=1.26.0
zeroconf>=0.131.0
opencv-contrib-python>=4.9.0.80
face-recognition>=1.3.0
